Moved fixtures to fixture.py
