+++ test/test_diffractometer_routes.py
@@ -1,10 +1,5 @@
-import sys
-import pytest
 import json
 import random
-
-from mxcube3 import server
-
 
 # Python 2 and 3 compatibility
 try:
@@ -12,20 +7,7 @@ try:
 except:
     unicode = str
 
-sys.path.append("./")
-
-
-@pytest.fixture
-def client():
-    server.config["TESTING"] = True
-
-    client = server.test_client()
-
-    data = json.dumps({"proposal": "idtest0", "password": "sUpErSaFe"})
-
-    client.post("/mxcube/api/v0.1/login", data=data, content_type="application/json")
-
-    yield client
+from fixture import client
 
 
 def test_get_phase_list(client):
